refactor: relocate ffmpeg bridge to block-utils (app + cli share it)
Plan Task 10. Move the host-agnostic ffmpeg bridge — FfmpegService trait,
ExecArgs/ExecResult/FfmpegError, and FfmpegBlock — from the browser-first app
crate into the native-compilable gizza-ai-block-utils crate, so the app
(browser) and the upcoming CLI (native) register the IDENTICAL FfmpegBlock with
their own FfmpegService impl. BrowserFfmpegService stays in the app crate (its
#[wasm_bindgen(module="/js/ffmpeg.js")] path is crate-relative) and now impls the
relocated trait; src/ffmpeg.rs re-exports the moved types so app call sites are
unchanged. Deletes src/blocks/ffmpeg.rs. Pure move — no behavior change.

ExecArgs/ExecResult vs FfmpegReq/FfmpegResp duplication left as-is (separate
cleanup, out of scope).

Verified: block-utils builds native + wasm32; APP still builds wasm32 (regression
check); block-utils 34 tests pass; CLI still builds.
