+++ block-utils/Cargo.toml
@@ -11,10 +11,11 @@ description = "Shared helpers for gizza-ai skill blocks (URL parsing, mime table
 license = "MIT"
 
 [dependencies]
+async-trait = "0.1"
 serde = { version = "1", features = ["derive"] }
+serde_json = "1"
 thiserror = "1"
 wafer-sdk = { git = "https://github.com/wafer-run/wafer-run.git", branch = "main" }
 wafer-block = { git = "https://github.com/wafer-run/wafer-run.git", branch = "main" }
 
 [dev-dependencies]
-serde_json = "1"
